+++ examples/demo/config.js
@@ -1,636 +1,112 @@
 import React from 'react';
-import { Widgets, Operators } from 'react-awesome-query-builder';
-
-const {
-    TextWidget,
-    NumberWidget,
-    SliderWidget,
-    RangeWidget,
-    SelectWidget,
-    MultiSelectWidget,
-    DateWidget,
-    BooleanWidget,
-    TimeWidget,
-    DateTimeWidget,
-    ValueFieldWidget
-} = Widgets;
-
-const { ProximityOperator } = Operators;
-import moment from 'moment';
+import merge from 'lodash/merge';
+import { Widgets, Operators, BasicConfig } from 'react-awesome-query-builder';
 import en_US from 'antd/lib/locale-provider/en_US';
 import ru_RU from 'antd/lib/locale-provider/ru_RU';
 
 const conjunctions = {
-    AND: {
-        label: 'And',
-        mongoConj: '$and',
-        reversedConj: 'OR',
-        formatConj: (children, conj, not, isForDisplay) => {
-            return children.size > 1 ?
-                (not ? "NOT " : "") + '(' + children.join(' ' + (isForDisplay ? "AND" : "&&") + ' ') + ')'
-                : (not ? "NOT (" : "") + children.first() + (not ? ")" : "");
-        },
-    },
-    OR: {
-        label: 'Or',
-        mongoConj: '$or',
-        reversedConj: 'AND',
-        formatConj: (children, conj, not, isForDisplay) => {
-            return children.size > 1 ?
-                (not ? "NOT " : "") + '(' + children.join(' ' + (isForDisplay ? "OR" : "||") + ' ') + ')'
-                : (not ? "NOT (" : "") + children.first() + (not ? ")" : "");
-        },
-    },
+    ...BasicConfig.conjunctions
 };
 
 const operators = {
-    equal: {
-        label: '==',
-        labelForFormat: '==',
-        reversedOp: 'not_equal',
-        mongoFormatOp: (field, op, value) => ({ [field]: { '$eq': value } }),
-    },
-    not_equal: {
-        label: '!=',
-        labelForFormat: '!=',
-        reversedOp: 'equal',
-        mongoFormatOp: (field, op, value) => ({ [field]: { '$ne': value } }),
-    },
-    less: {
-        label: '<',
-        labelForFormat: '<',
-        reversedOp: 'greater_or_equal',
-        mongoFormatOp: (field, op, value) => ({ [field]: { '$lt': value } }),
-    },
-    less_or_equal: {
-        label: '<=',
-        labelForFormat: '<=',
-        reversedOp: 'greater',
-        mongoFormatOp: (field, op, value) => ({ [field]: { '$lte': value } }),
-    },
-    greater: {
-        label: '>',
-        labelForFormat: '>',
-        reversedOp: 'less_or_equal',
-        mongoFormatOp: (field, op, value) => ({ [field]: { '$gt': value } }),
-    },
-    greater_or_equal: {
-        label: '>=',
-        labelForFormat: '>=',
-        reversedOp: 'less',
-        mongoFormatOp: (field, op, value) => ({ [field]: { '$gte': value } }),
-    },
+    ...BasicConfig.operators,
+    // examples of  overriding
     between: {
-        label: 'Between',
-        labelForFormat: 'BETWEEN',
-        cardinality: 2,
-        formatOp: (field, op, values, valueSrcs, valueTypes, opDef, operatorOptions, isForDisplay) => {
-            let valFrom = values.first();
-            let valTo = values.get(1);
-            if (isForDisplay)
-                return `${field} >= ${valFrom} AND ${field} <= ${valTo}`;
-            else
-                return `${field} >= ${valFrom} && ${field} <= ${valTo}`;
-        },
-        mongoFormatOp: (field, op, values) => ({ [field]: { '$gte': values[0], '$lte': values[1] } }),
+        ...BasicConfig.operators.between,
         valueLabels: [
             'Value from',
             'Value to'
         ],
         textSeparators: [
-            null,
-            'and'
+            'from',
+            'to'
         ],
-        reversedOp: 'not_between',
-    },
-    not_between: {
-        label: 'Not between',
-        labelForFormat: 'NOT BETWEEN',
-        cardinality: 2,
-        mongoFormatOp: (field, op, values) => ({ [field]: { '$not': { '$gte': values[0], '$lte': values[1] } } }),
-        valueLabels: [
-            'Value from',
-            'Value to'
-        ],
-        textSeparators: [
-            null,
-            'and'
-        ],
-        reversedOp: 'between',
-    },
-    range_between: {
-        label: 'Between',
-        labelForFormat: 'BETWEEN',
-        cardinality: 2,
-        isSpecialRange: true, // to show 1 range widget instead of 2
-        formatOp: (field, op, values, valueSrcs, valueTypes, opDef, operatorOptions, isForDisplay) => {
-            let valFrom = values.first();
-            let valTo = values.get(1);
-            if (isForDisplay)
-                return `${field} >= ${valFrom} AND ${field} <= ${valTo}`;
-            else
-                return `${field} >= ${valFrom} && ${field} <= ${valTo}`;
-        },
-        mongoFormatOp: (field, op, values) => ({ [field]: { '$gte': values[0], '$lte': values[1] } }),
-        valueLabels: [
-            'Value from',
-            'Value to'
-        ],
-        textSeparators: [
-            null,
-            'and'
-        ],
-        reversedOp: 'range_not_between',
-    },
-    range_not_between: {
-        label: 'Not between',
-        labelForFormat: 'NOT BETWEEN',
-        cardinality: 2,
-        isSpecialRange: true, // to show 1 range widget instead of 2
-        mongoFormatOp: (field, op, values) => ({ [field]: { '$not': { '$gte': values[0], '$lte': values[1] } } }),
-        valueLabels: [
-            'Value from',
-            'Value to'
-        ],
-        textSeparators: [
-            null,
-            'and'
-        ],
-        reversedOp: 'range_between',
-    },
-    is_empty: {
-        isUnary: true,
-        label: 'Is Empty',
-        labelForFormat: 'IS EMPTY',
-        cardinality: 0,
-        reversedOp: 'is_not_empty',
-        formatOp: (field, op, value, valueSrc, valueType, opDef, operatorOptions, isForDisplay) => {
-            return isForDisplay ? `${field} IS EMPTY` : `!${field}`;
-        },
-        mongoFormatOp: (field, op) => ({ [field]: { '$exists': false } }),
-    },
-    is_not_empty: {
-        isUnary: true,
-        label: 'Is not empty',
-        labelForFormat: 'IS NOT EMPTY',
-        cardinality: 0,
-        reversedOp: 'is_empty',
-        formatOp: (field, op, value, valueSrc, valueType, opDef, operatorOptions, isForDisplay) => {
-            return isForDisplay ? `${field} IS NOT EMPTY` : `!!${field}`;
-        },
-        mongoFormatOp: (field, op) => ({ [field]: { '$exists': true } }),
-    },
-    select_equals: {
-        label: '==',
-        labelForFormat: '==',
-        formatOp: (field, op, value, valueSrc, valueType, opDef, operatorOptions, isForDisplay) => {
-            return `${field} == ${value}`;
-        },
-        mongoFormatOp: (field, op, value) => ({ [field]: { '$eq': value } }),
-        reversedOp: 'select_not_equals',
-    },
-    select_not_equals: {
-        label: '!=',
-        labelForFormat: '!=',
-        formatOp: (field, op, value, valueSrc, valueType, opDef, operatorOptions, isForDisplay) => {
-            return `${field} != ${value}`;
-        },
-        mongoFormatOp: (field, op, value) => ({ [field]: { '$ne': value } }),
-        reversedOp: 'select_equals',
-    },
-    select_any_in: {
-        label: 'Any in',
-        labelForFormat: 'IN',
-        formatOp: (field, op, values, valueSrc, valueType, opDef, operatorOptions, isForDisplay) => {
-            if (valueSrc == 'value')
-                return `${field} IN (${values.join(', ')})`;
-            else
-                return `${field} IN (${values})`;
-        },
-        mongoFormatOp: (field, op, values) => ({ [field]: { '$in': values } }),
-        reversedOp: 'select_not_any_in',
-    },
-    select_not_any_in: {
-        label: 'Not in',
-        labelForFormat: 'NOT IN',
-        formatOp: (field, op, values, valueSrc, valueType, opDef, operatorOptions, isForDisplay) => {
-            if (valueSrc == 'value')
-                return `${field} NOT IN (${values.join(', ')})`;
-            else
-                return `${field} NOT IN (${values})`;
-        },
-        mongoFormatOp: (field, op, values) => ({ [field]: { '$nin': values } }),
-        reversedOp: 'select_any_in',
-    },
-    multiselect_equals: {
-        label: 'Equals',
-        labelForFormat: '==',
-        formatOp: (field, op, values, valueSrc, valueType, opDef, operatorOptions, isForDisplay) => {
-            if (valueSrc == 'value')
-                return `${field} == [${values.join(', ')}]`;
-            else
-                return `${field} == ${values}`;
-        },
-        mongoFormatOp: (field, op, values) => ({ [field]: { '$eq': values } }),
-        reversedOp: 'multiselect_not_equals',
-    },
-    multiselect_not_equals: {
-        label: 'Not equals',
-        labelForFormat: '!=',
-        formatOp: (field, op, values, valueSrc, valueType, opDef, operatorOptions, isForDisplay) => {
-            if (valueSrc == 'value')
-                return `${field} != [${values.join(', ')}]`;
-            else
-                return `${field} != ${values}`;
-        },
-        mongoFormatOp: (field, op, values) => ({ [field]: { '$ne': values } }),
-        reversedOp: 'multiselect_equals',
     },
     proximity: {
-        label: 'Proximity search',
-        cardinality: 2,
+        ...BasicConfig.operators.proximity,
         valueLabels: [
             { label: 'Word 1', placeholder: 'Enter first word' },
-            'Word 2'
+            { label: 'Word 2', placeholder: 'Enter second word' },
         ],
         textSeparators: [
             //'Word 1',
             //'Word 2'
         ],
-        formatOp: (field, op, values, valueSrc, valueType, opDef, operatorOptions, isForDisplay) => {
-            let val1 = values.first();
-            let val2 = values.get(1);
-            return `${field} ${val1} NEAR/${operatorOptions.get('proximity')} ${val2}`;
-        },
-        mongoFormatOp: (field, op, values) => (undefined),
         options: {
-            optionLabel: "Near",
-            optionTextBefore: "Near",
-            optionPlaceholder: "Select words between",
-            factory: (props) => <ProximityOperator {...props} />,
-            defaults: {
-                proximity: 2
-            }
+            ...BasicConfig.operators.proximity.options,
+            optionLabel: "Near", // label on top of "near" selectbox (for config.settings.showLabels==true)
+            optionTextBefore: "Near", // label before "near" selectbox (for config.settings.showLabels==false)
+            optionPlaceholder: "Select words between", // placeholder for "near" selectbox
+            minProximity: 2,
+            maxProximity: 10,
+            defaultProximity: 2,
+            customProps: {}
         }
     },
 };
 
 const widgets = {
+    ...BasicConfig.widgets,
+    // examples of  overriding
     text: {
-        type: "text",
-        valueSrc: 'value',
-        factory: (props) => <TextWidget {...props} />,
-        formatValue: (val, fieldDef, wgtDef, isForDisplay) => {
-            return isForDisplay ? '"' + val + '"' : JSON.stringify(val);
-        },
+        ...BasicConfig.widgets.text,
         validateValue: (val, fieldDef) => {
             return (val.length < 10);
         },
     },
-    number: {
-        type: "number",
-        valueSrc: 'value',
-        factory: (props) => <NumberWidget {...props} />,
-        valueLabel: "Number",
-        valuePlaceholder: "Enter number",
-        formatValue: (val, fieldDef, wgtDef, isForDisplay) => {
-            return isForDisplay ? val : JSON.stringify(val);
-        },
-    },
     slider: {
-        type: "number",
-        valueSrc: 'value',
-        factory: (props) => <SliderWidget {...props} />,
-        valueLabel: "Slider",
-        valuePlaceholder: "Move Slider",
-        formatValue: (val, fieldDef, wgtDef, isForDisplay) => {
-            return isForDisplay ? val : JSON.stringify(val);
-        },
+        ...BasicConfig.widgets.slider,
         customProps: {
             width: '300px'
         }
     },
     rangeslider: {
-        type: "number",
-        valueSrc: 'value',
-        factory: (props) => <RangeWidget {...props} />,
-        valueLabel: "Range",
-        valuePlaceholder: "Select Range",
-        formatValue: (val, fieldDef, wgtDef, isForDisplay) => {
-            return isForDisplay ? val : JSON.stringify(val);
-        },
+        ...BasicConfig.widgets.rangeslider,
         customProps: {
             width: '300px'
-        },
-        singleWidget: 'slider',
-    },
-    select: {
-        type: "select",
-        factory: (props) => <SelectWidget {...props} />,
-        formatValue: (val, fieldDef, wgtDef, isForDisplay) => {
-            let valLabel = fieldDef.listValues[val];
-            return isForDisplay ? '"' + valLabel + '"' : JSON.stringify(val);
-        },
-    },
-    multiselect: {
-        type: "multiselect",
-        valueSrc: 'value',
-        factory: (props) => <MultiSelectWidget {...props} />,
-        formatValue: (vals, fieldDef, wgtDef, isForDisplay) => {
-            let valsLabels = vals.map(v => fieldDef.listValues[v]);
-            return isForDisplay ? valsLabels.map(v => '"' + v + '"') : vals.map(v => JSON.stringify(v));
-        },
+        }
     },
     date: {
-        type: "date",
-        valueSrc: 'value',
-        factory: (props) => <DateWidget {...props} />,
+        ...BasicConfig.widgets.date,
         dateFormat: 'DD.MM.YYYY',
         valueFormat: 'YYYY-MM-DD',
-        formatValue: (val, fieldDef, wgtDef, isForDisplay) => {
-            let dateVal = moment(val, wgtDef.valueFormat);
-            return isForDisplay ? '"' + dateVal.format(wgtDef.dateFormat) + '"' : JSON.stringify(val);
-        },
     },
     time: {
-        type: "time",
-        valueSrc: 'value',
-        factory: (props) => <TimeWidget {...props} />,
+        ...BasicConfig.widgets.time,
         timeFormat: 'HH:mm',
         valueFormat: 'HH:mm:ss',
-        formatValue: (val, fieldDef, wgtDef, isForDisplay) => {
-            let dateVal = moment(val, wgtDef.valueFormat);
-            return isForDisplay ? '"' + dateVal.format(wgtDef.timeFormat) + '"' : JSON.stringify(val);
-        },
     },
     datetime: {
-        type: "datetime",
-        valueSrc: 'value',
-        factory: (props) => <DateTimeWidget {...props} />,
+        ...BasicConfig.widgets.datetime,
         timeFormat: 'HH:mm',
         dateFormat: 'DD.MM.YYYY',
         valueFormat: 'YYYY-MM-DD HH:mm:ss',
-        formatValue: (val, fieldDef, wgtDef, isForDisplay) => {
-            let dateVal = moment(val, wgtDef.valueFormat);
-            return isForDisplay ? '"' + dateVal.format(wgtDef.dateFormat + ' ' + wgtDef.timeFormat) + '"' : JSON.stringify(val);
-        },
     },
-    boolean: {
-        type: "boolean",
-        valueSrc: 'value',
-        factory: (props) => <BooleanWidget {...props} />,
-        labelYes: "Yes",
-        labelNo: "No ",
-        formatValue: (val, fieldDef, wgtDef, isForDisplay) => {
-            return isForDisplay ? (val ? "Yes" : "No") : JSON.stringify(!!val);
-        },
-        defaultValue: false,
-    },
-    field: {
-        valueSrc: 'field',
-        factory: (props) => <ValueFieldWidget {...props} />,
-        formatValue: (val, fieldDef, wgtDef, isForDisplay, valFieldDef) => {
-            return isForDisplay ? (valFieldDef.label || val) : val;
-        },
-        valueLabel: "Field to compare",
-        valuePlaceholder: "Select field to compare",
-        customProps: {
-            showSearch: true
-        }
-    }
 };
+
 
 const types = {
-    text: {
-        widgets: {
-            text: {
-                defaultOperator: 'is_empty',
-                operators: [
-                    'equal',
-                    'not_equal',
-                    "is_empty",
-                    "is_not_empty",
-                    'proximity'
-                ],
-                widgetProps: {
-                    formatValue: (val, fieldDef, wgtDef, isForDisplay) => (JSON.stringify(val)),
-                    valueLabel: "Text",
-                    valuePlaceholder: "Enter text",
-                }
-            },
-            field: {
-                operators: [
-                    'equal',
-                    'not_equal',
-                    //note that unary ops will be excluded anyway, see getWidgetsForFieldOp()
-                    //"is_empty",
-                    //"is_not_empty",
-                    'proximity'
-                ],
-            }
-        },
-    },
-    number: {
-        mainWidget: "number",
-        valueSources: ['value', 'field'],
-        defaultOperator: 'equal',
-        widgets: {
-            number: {
-                operators: [
-                    "equal",
-                    "not_equal",
-                    "less",
-                    "less_or_equal",
-                    "greater",
-                    "greater_or_equal",
-                    "between",
-                    "not_between",
-                    "is_empty",
-                    "is_not_empty",
-                ],
-                widgetProps: {
-                    valueLabel: "Number2",
-                    valuePlaceholder: "Enter number2",
-                }
-            },
-            slider: {
-                operators: [
-                    "equal",
-                    "not_equal",
-                    "less",
-                    "less_or_equal",
-                    "greater",
-                    "greater_or_equal",
-                ],
-                widgetProps: {
-                    valueLabel: "Slider",
-                    valuePlaceholder: "Move slider",
-                    customProps: {
-                        width: '200px'
-                    }
-                }
-            },
-            rangeslider: {
-                operators: [
-                    "range_between",
-                    "range_not_between",
-                ],
-                widgetProps: {
-                    valueLabel: "Range",
-                    valuePlaceholder: "Select range",
-                    customProps: {
-                        width: '300px'
-                    }
-                }
-            }
-        },
-    },
-    date: {
-        widgets: {
-            date: {
-                operators: [
-                    "equal",
-                    "not_equal",
-                    "less",
-                    "less_or_equal",
-                    "greater",
-                    "greater_or_equal",
-                    "between",
-                    "not_between",
-                    "is_empty",
-                    "is_not_empty",
-                ]
-            }
-        },
-    },
-    time: {
-        widgets: {
-            time: {
-                operators: [
-                    "equal",
-                    "not_equal",
-                    "less",
-                    "less_or_equal",
-                    "greater",
-                    "greater_or_equal",
-                    "between",
-                    "not_between",
-                    "is_empty",
-                    "is_not_empty",
-                ]
-            }
-        },
-    },
-    datetime: {
-        widgets: {
-            datetime: {
-                operators: [
-                    "equal",
-                    "not_equal",
-                    "less",
-                    "less_or_equal",
-                    "greater",
-                    "greater_or_equal",
-                    "between",
-                    "not_between",
-                    "is_empty",
-                    "is_not_empty",
-                ],
-                opProps: {
-                    between: {
-                        valueLabels: [
-                            { label: 'Date from', placeholder: 'Enrer datetime from' },
-                            { label: 'Date to', placeholder: 'Enter datetime to' },
-                        ],
-                    },
-                },
-                widgetProps: {
-                    timeFormat: 'HH:mm',
-                    dateFormat: 'YYYY-MM-DD',
-                    valueFormat: 'YYYY-MM-DD HH:mm',
-                }
-            }
-        },
-    },
-    select: {
-        mainWidget: "select",
-        widgets: {
-            select: {
-                operators: [
-                    'select_equals',
-                    'select_not_equals'
-                ],
-                widgetProps: {
-                    customProps: {
-                        showSearch: true
-                    }
-                },
-            },
-            multiselect: {
-                operators: [
-                    'select_any_in',
-                    'select_not_any_in'
-                ],
-                widgetProps: {
-                },
-            },
-        },
-    },
-    multiselect: {
-        widgets: {
-            multiselect: {
-                operators: [
-                    'multiselect_equals',
-                    'multiselect_not_equals',
-                ]
-            }
-        },
-    },
-    boolean: {
+    ...BasicConfig.types,
+    // examples of  overriding
+    boolean: merge(BasicConfig.types.boolean, {
         widgets: {
             boolean: {
-                operators: [
-                    "equal",
-                ],
                 widgetProps: {
-                    //you can enable this if you don't use fields as value sources
-                    //hideOperator: true,
-                    //operatorInlineLabel: "is",
+                    hideOperator: true,
+                    operatorInlineLabel: "is",
                 }
             },
-            field: {
-                operators: [
-                    "equal",
-                    "not_equal",
-                ],
-            }
         },
-    },
+    }),
 };
 
-const settings = {
+
+const localeSettings = {
     locale: {
-        short: 'en',
-        full: 'en-US',
-        antd: en_US,
+        short: 'ru',
+        full: 'ru-RU',
+        antd: ru_RU,
     },
-    maxLabelsLength: 50,
-    hideConjForOne: true,
-    renderSize: 'small',
-    renderConjsAsRadios: false,
-    renderFieldAndOpAsDropdown: false,
-    customFieldSelectProps: {
-        showSearch: true
-    },
-    groupActionsPosition: 'topRight', // oneOf [topLeft, topCenter, topRight, bottomLeft, bottomCenter, bottomRight]
-    setOpOnChangeField: ['keep', 'default'], // 'default' (default if present), 'keep' (keep prev from last field), 'first', 'none'
-    clearValueOnChangeField: false, //false - if prev & next fields have same type (widget), keep
-    clearValueOnChangeOp: false,
-    setDefaultFieldAndOp: false,
-    maxNesting: 10,
-    fieldSeparator: '.',
-    fieldSeparatorDisplay: '->',
-    showLabels: false,
     valueLabel: "Value",
     valuePlaceholder: "Value",
     fieldLabel: "Field",
@@ -640,10 +116,9 @@ const settings = {
     deleteLabel: null,
     addGroupLabel: "Add group",
     addRuleLabel: "Add rule",
-    immutableGroupsMode: false,
-    canReorder: true,
-    canRegroup: true,
+    delGroupLabel: null,
     notLabel: "Not",
+    valueSourcesPopupTitle: "Select value source",
     removeRuleConfirmOptions: {
         title: 'Are you sure delete this rule?',
         okText: 'Yes',
@@ -654,21 +129,12 @@ const settings = {
         okText: 'Yes',
         okType: 'danger',
     },
-    showNot: true,
-    delGroupLabel: null,
-    canLeaveEmptyGroup: true, //after deletion
-    formatReverse: (q, operator, reversedOp, operatorDefinition, revOperatorDefinition, isForDisplay) => {
-        if (isForDisplay)
-            return "NOT(" + q + ")";
-        else
-            return "!(" + q + ")";
-    },
-    formatField: (field, parts, label2, fieldDefinition, config, isForDisplay) => {
-        if (isForDisplay)
-            return label2;
-        else
-            return field;
-    },
+  };
+
+const settings = {
+    ...BasicConfig.settings,
+    ...localeSettings,
+
     valueSourcesInfo: {
         value: {
             label: "Value"
@@ -678,12 +144,15 @@ const settings = {
             widget: "field",
         }
     },
-    valueSourcesPopupTitle: "Select value source",
-    canCompareFieldWithField: (leftField, leftFieldConfig, rightField, rightFieldConfig) => {
-        //for type == 'select'/'multiselect' you can check listValues
-        return true;
-    },
+    // canReorder: true,
+    // canRegroup: true,
+    // showNot: true,
+    // showLabels: true,
+    maxNesting: 3,
+    canLeaveEmptyGroup: true, //after deletion
 };
+
+//////////////////////////////////////////////////////////////////////
 
 const fields = {
     members: {
@@ -696,7 +165,7 @@ const fields = {
                 label2: 'MemberName', //only for menu's toggler
                 type: 'text',
                 tableName: 't1', // PR #18, PR #20
-                operators: ['equal'],
+                //operators: ['equal'],
             },
             prox1: {
                 label: 'prox',
@@ -875,6 +344,7 @@ const fields = {
         type: 'boolean',
     },
     expecting: {
+        valueSources: ['value'],
         label: 'Expecting',
         type: 'boolean',
     },
@@ -882,10 +352,10 @@ const fields = {
 
 export default {
     conjunctions,
-    fields,
-    types,
     operators,
     widgets,
+    types,
     settings,
+    fields: fields,
 };
