+++ examples/demo/demo.js
@@ -1,21 +1,21 @@
 import React, {Component} from 'react';
-import {Query, Builder, Utils} from 'react-awesome-query-builder';
-const {queryBuilderFormat, queryString, mongodbFormat, getTree, loadTree, uuid} = Utils;
+import {Query, Builder, BasicConfig, Utils} from 'react-awesome-query-builder';
+const {queryBuilderFormat, queryString, mongodbFormat, getTree, checkTree, loadTree, uuid} = Utils;
 const stringify = require('json-stringify-safe');
 import throttle from 'lodash/throttle';
 import Immutable from 'immutable';
-window.Immutable = Immutable
+window.Immutable = Immutable;
+const preStyle = { backgroundColor: 'darkgrey', margin: '10px', padding: '10px' };
 
 import loadedConfig from './config';
 import loadedInitValue from './init_value';
 const emptyInitValue = {"id": uuid(), "type": "group"};
 const initValue = loadedInitValue && Object.keys(loadedInitValue).length > 0 ? loadedInitValue : emptyInitValue;
-const preStyle = { backgroundColor: 'darkgrey', margin: '10px', padding: '10px' };
 
 
 export default class DemoQueryBuilder extends Component {
     state = {
-      tree: loadTree(initValue),
+      tree: checkTree(loadTree(initValue), loadedConfig),
       config: loadedConfig
     };
