+++ modules/config/default.js
@@ -1,0 +1,45 @@
+import en_US from 'antd/lib/locale-provider/en_US';
+
+export const settings = {
+  formatField: (field, parts, label2, fieldDefinition, config, isForDisplay) => {
+    if (isForDisplay)
+        return label2;
+    else
+        return field;
+  },
+
+  valueSourcesInfo: {
+      value: {},
+  },
+  fieldSeparator: '.',
+  fieldSeparatorDisplay: '->',
+  renderSize: "small",
+  maxLabelsLength: 100,
+  hideConjForOne: true,
+  canReorder: true,
+  canRegroup: true,
+  showNot: true,
+  groupActionsPosition: 'topRight', // oneOf [topLeft, topCenter, topRight, bottomLeft, bottomCenter, bottomRight]
+  setOpOnChangeField: ['keep', 'default'], // 'default' (default if present), 'keep' (keep prev from last field), 'first', 'none'
+
+  // localization
+  locale: {
+    short: 'en',
+    full: 'en-US',
+    antd: en_US,
+  },
+  valueLabel: "Value",
+  valuePlaceholder: "Value",
+  fieldLabel: "Field",
+  operatorLabel: "Operator",
+  fieldPlaceholder: "Select field",
+  operatorPlaceholder: "Select operator",
+  deleteLabel: null,
+  addGroupLabel: "Add group",
+  addRuleLabel: "Add rule",
+  delGroupLabel: null,
+  notLabel: "Not",
+  valueSourcesPopupTitle: "Select value source",
+  removeRuleConfirmOptions: null,
+  removeGroupConfirmOptions: null,
+};
