+++ modules/index.js
@@ -1,8 +1,8 @@
 export { default as Query } from './components/Query';
 export { default as Builder } from './components/Builder';
-export { default as Preview } from './components/Preview';
 
 export * as Widgets from './components/widgets/index.js';
 export * as Operators from './components/operators';
 export * as Utils from './utils';
+export {default as BasicConfig} from './config/basic';
