+++ modules/utils/index.js
@@ -2,7 +2,7 @@
 export {queryBuilderFormat} from './queryBuilderFormat'
 export {mongodbFormat} from './mongodbFormat'
 export {queryString} from './queryString'
-export {getTree, loadTree} from './treeUtils'
+export {getTree, loadTree, checkTree} from './treeUtils'
 export {validateTree} from './validation'
 export {default as uuid} from './uuid';
 export {
